% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigo.R, R/complexity.R
\docType{package}
\name{bigo}
\alias{bigo}
\alias{bigo-package}
\alias{bigo}
\title{bigo: Analysis of Algorithmic Complexity}
\usage{
bigo(f, ..., num_runs = 1)
}
\arguments{
\item{f}{The function to be investigated.}

\item{...}{One or more numeric or integer vectors, passed as named
arguments, where the names are arguments of \code{f}.}

\item{num_runs}{The number of experiments over which to average results.}
}
\value{
A tibble of class \code{bigo} with containing the results of the
  runtime experiments
}
\description{
Simple tools for analyzing algorithmic complexity.

Analyze the algorithmic complexity of a function with respect to arguments.
}
\details{
This function returns an object of class \code{bigo}, which contains a
tibble of runtime results, the name of the function being measured, and
the number of runs over which the results are averaged.
}
\examples{
bigo(f = function(n) if (n < 3) 1 else { Recall(n - 1) + Recall(n - 2) },
     n = 1:15)
}
