% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigO.R
\name{BigO}
\alias{BigO}
\title{kaggle competition}
\usage{
BigO(training, test.real, outputCSV = TRUE)
}
\arguments{
\item{training}{data set for training}

\item{test}{data set for test}
}
\value{
a dataframe with predictied class on test data set
}
\description{
kaggle competition
}

